+++ src/consumer/component/components-list.js
@@ -299,9 +299,16 @@ export default class ComponentsList {
   /**
    * get called when the Consumer is available, shows also components from remote scopes
    */
-  async listScope(showRemoteVersion: boolean, includeNested: boolean): Promise<ListScopeResult[]> {
+  async listScope(
+    showRemoteVersion: boolean,
+    includeNested: boolean,
+    namespacesUsingWildcards?: string
+  ): Promise<ListScopeResult[]> {
     const components: ModelComponent[] = await this.getModelComponents();
-    const componentsSorted = ComponentsList.sortComponentsByName(components);
+    const componentsFilteredByWildcards = namespacesUsingWildcards
+      ? ComponentsList.filterComponentsByWildcard(components, namespacesUsingWildcards)
+      : components;
+    const componentsSorted = ComponentsList.sortComponentsByName(componentsFilteredByWildcards);
     const listScopeResults: ListScopeResult[] = componentsSorted.map((component: ModelComponent) => ({
       id: component.toBitIdWithLatestVersion(),
       deprecated: component.deprecated